adding all tokens for initial file.neo
